refactor TemplateCache and BaselineCache
